found a typo had %llu  should of had %lu
git-svn-id: http://svn.perl.org/modules/dbd-oracle/trunk@13769 50811bd7-b8ce-0310-adc1-d9db26280581

+++ ocitrace.h
@@ -222,8 +222,8 @@
 	stat = OCINlsEnvironmentVariableGet(	valp, size, item, charset, rsizep ); \
 	(DBD_OCI_TRACEON) \
 	?	PerlIO_printf(DBD_OCI_TRACEFP,\
-		 "%sNlsEnvironmentVariableGet(%d,%llu,%d,%d,%llu)=%s\n",\
-		 OciTp, *valp, (unsigned long long)size, item, charset, (unsigned long long)*rsizep, oci_status_name(stat)),stat \
+		 "%sNlsEnvironmentVariableGet(%d,%lu,%d,%d,%lu)=%s\n",\
+		 OciTp, *valp, (unsigned long)size, item, charset, (unsigned long)*rsizep, oci_status_name(stat)),stat \
 	: stat
 
 /* added by lab */
